requirements.txt
agent-framework==1.0.0b251114
azure-identity
pydantic>=2.0